Protect API with auth and add ownsership to expenses and categories

+++ src/server/router/expenses.ts
@@ -1,12 +1,16 @@
-import { createRouter } from './context';
+import { createProtectedRouter } from './protected-router';
 
 import { Expense, UpdateCategoriesSchema } from '@schemas/Expense';
 import { z } from 'zod';
+import { TRPCError } from '@trpc/server';
 
-export const expensesRouter = createRouter()
+export const expensesRouter = createProtectedRouter()
   .query('allExpenses', {
     async resolve({ ctx }) {
       return await ctx.prisma.expense.findMany({
+        where: {
+          userId: ctx.session.user.id,
+        },
         include: {
           categories: true,
         },
@@ -16,10 +20,15 @@ export const expensesRouter = createRouter()
   .mutation('createExpense', {
     input: Expense,
     async resolve({ ctx, input }) {
+      if (!ctx.session.user.id) {
+        throw new TRPCError({ code: 'FORBIDDEN' });
+      }
+
       const { name, amount, description, categories } = input;
 
       return await ctx.prisma.expense.create({
         data: {
+          userId: ctx.session.user.id,
           name,
           amount,
           description,
@@ -36,6 +45,21 @@ export const expensesRouter = createRouter()
   .mutation('deleteExpense', {
     input: z.object({ id: z.string() }),
     async resolve({ ctx, input }) {
+      if (!ctx.session.user.id) {
+        throw new TRPCError({ code: 'FORBIDDEN' });
+      }
+
+      const expense = await ctx.prisma.expense.findFirst({
+        where: {
+          id: input.id,
+          userId: ctx.session.user.id,
+        },
+      });
+
+      if (!expense) {
+        throw new TRPCError({ code: 'NOT_FOUND' });
+      }
+
       return await ctx.prisma.expense.delete({
         where: { id: input.id },
       });
@@ -44,7 +68,22 @@ export const expensesRouter = createRouter()
   .mutation('udpateCategories', {
     input: UpdateCategoriesSchema,
     async resolve({ ctx, input }) {
+      if (!ctx.session.user.id) {
+        throw new TRPCError({ code: 'FORBIDDEN' });
+      }
+
       const { id, categories } = input;
+
+      const expense = await ctx.prisma.expense.findFirst({
+        where: {
+          id,
+          userId: ctx.session.user.id,
+        },
+      });
+
+      if (!expense) {
+        throw new TRPCError({ code: 'NOT_FOUND' });
+      }
 
       return await ctx.prisma.expense.update({
         where: { id },
